# Efficient version of egraph.add_expr and extractor.find_best