--- a/src/extract.rs
+++ b/src/extract.rs
@@ -201,7 +201,12 @@ where
         (cost, expr)
     }
 
-    fn find_best_rec(&mut self, expr: &mut RecExpr<L>, eclass: Id, added_memo: &mut HashMap<Id, Id>) -> (Id, CF::Cost) {
+    fn find_best_rec(
+        &mut self,
+        expr: &mut RecExpr<L>,
+        eclass: Id,
+        added_memo: &mut HashMap<Id, Id>,
+    ) -> (Id, CF::Cost) {
         let id = self.egraph.find(eclass);
         let (best_cost, best_node) = match self.costs.get(&id) {
             Some(result) => result.clone(),
@@ -211,11 +216,12 @@ where
         match added_memo.get(&id) {
             Some(id_expr) => (*id_expr, best_cost),
             None => {
-                let node = best_node.map_children(|child| self.find_best_rec(expr, child, added_memo).0);
+                let node =
+                    best_node.map_children(|child| self.find_best_rec(expr, child, added_memo).0);
                 let id_expr = expr.add(node);
                 assert!(added_memo.insert(id, id_expr).is_none());
                 (id_expr, best_cost)
-            },
+            }
         }
     }