Add memoization to egraph.add_expr and extractor.find_best. On larger graphs, this gives a lot of speed up (changing from exponential to linear time). 

Run time comparison, on benchnet from https://github.com/yycdavid/tamago (~200 nodes):
-- add_expr: 
old version 1747s
new version 2.26s
-- find_best:
old version: 1177s
new version: 0.00123s